fix(experiments): address PR #604 review comments
1. Split src/experiments/diagnostics.py into a package (CODE.md: one class
   per file). New layout:
     src/experiments/diagnostics/
       __init__.py      — re-exports preserve the flat import path
       stats.py         — DistributionStats
       hit_rate.py      — HitRate
       report.py        — DiagnosticReport (+ _pct helper with docstring)
       runner.py        — SignalDiagnostic + DEFAULT_HORIZONS

2. Add docstring to _pct helper (CODE.md: always include a docstring).

3. Fix DataFrame column-validation bug in SignalDiagnostic._walk_bars.
   The previous code did ``df["close"].to_numpy() if hasattr(df, "to_numpy")``
   which raised a bare KeyError when the column was missing — the helpful
   ValueError that followed was unreachable. Now validates
   ``"close" in df.columns`` BEFORE indexing.

4. Split compound asserts per CODE.md:
     - test_diagnostics.py: ``assert A and B`` (to_dict check) → two asserts
     - test_diagnostics.py: ``assert A and B and C`` (BUY/SELL/HOLD) → three
     - test_reporter_edges.py: ``assert A or B`` (CSV line endings) →
       normalize ``\r\n`` → ``\n`` before a single assertion.

5. Switch MagicMock to create_autospec in test_suite_runner.py
   (CODE.md: use autospec to catch signature mismatches).
   Added ``_mock_runner()`` helper returning
   ``create_autospec(ExperimentRunner, instance=True)``; replaced all four
   ``ExperimentSuiteRunner(runner=MagicMock())`` sites.

All 193 experiment unit tests still pass. Ruff + black clean.
